backend/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
requests==2.32.3
urllib3==2.2.2
websockets==12.0
loguru==0.7.2
streamcontroller-plugin-tools>=2.0.1
pillow==10.3.0
psutil==6.0.0